[MINOR] fix several printf formats and missing arguments
Last patch revealed a number of mistakes in printf-like calls, mostly int/long
mismatches, and a few missing arguments.

(cherry-picked from commit 1772ece0252fa517ace6c554ee77ca09c3dda7e8)
(cherry-picked from commit 8169959df85d50356a0beec0c74e23dcae1d7c32)

+++ src/dumpstats.c
@@ -882,7 +882,7 @@ int stats_dump_proxy(struct session *s, struct proxy *px, struct uri_auth *uri)
 					chunk_printf(&msg, sizeof(trash),
 					     "%d,%d,%d,%d,",
 					     sv->failed_checks, sv->down_trans,
-					     now.tv_sec - sv->last_change, srv_downtime(sv));
+					     (int)(now.tv_sec - sv->last_change), srv_downtime(sv));
 				else
 					chunk_printf(&msg, sizeof(trash),
 					     ",,,,");
@@ -1001,7 +1001,7 @@ int stats_dump_proxy(struct session *s, struct proxy *px, struct uri_auth *uri)
 				     (px->lbprm.tot_weight > 0 || !px->srv) ? "UP" : "DOWN",
 				     (px->lbprm.tot_weight * px->lbprm.wmult + px->lbprm.wdiv - 1) / px->lbprm.wdiv,
 				     px->srv_act, px->srv_bck,
-				     px->down_trans, now.tv_sec - px->last_change,
+				     px->down_trans, (int)(now.tv_sec - px->last_change),
 				     px->srv?be_downtime(px):0,
 				     relative_pid, px->uuid,
 				     px->cum_lbconn);
